pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voicetry"
version = "1.0.0"
description = "Futuristic voice AI terminal with real-time visualization"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "VoiceTry"}
]
keywords = ["voice", "ai", "tts", "stt", "terminal", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "soundfile>=0.12.1",
    "openai-whisper>=20231117",
    "openai>=1.0.0",
    "kokoro>=0.9.4",
    "misaki[en]>=0.0.9",
    "rich>=13.7.0",
    "readchar>=4.0.5",
    "python-dotenv>=1.0.0",
]

[project.scripts]
voicetry = "src.main:main"

[project.urls]
Homepage = "https://github.com/user/voicetry"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "src/**"]
